cast usgs_df values to single

+++ src/python_routing_v02/compute_nhd_routing_SingleSeg_v02.py
@@ -590,7 +590,7 @@ def compute_nhd_routing_v02(
                         q0_sub.values,
                         qlat_sub.values,
                         {},
-                        usgs_df_sub.values,
+                        np.single(usgs_df_sub.values),
                         assume_short_ts,
                     )
                 )
@@ -620,7 +620,7 @@ def compute_nhd_routing_v02(
                     q0_sub.values,
                     qlat_sub.values,
                     {},
-                    usgs_df_sub.values,
+                    np.single(usgs_df_sub.values),
                     assume_short_ts,
                 )
             )
